Reorganize packages (#106)
* Reorg: Move utils to subpackage
* Reorg: Move dtos to subpackage
* Reorg: Move test utils
