+++ src/main/java/com/ginsberg/gatherers4j/ZipWithGatherer.java
@@ -16,6 +16,7 @@
 
 package com.ginsberg.gatherers4j;
 
+import com.ginsberg.gatherers4j.dto.Pair;
 import org.jspecify.annotations.Nullable;
 
 import java.util.Arrays;
@@ -26,7 +27,7 @@ import java.util.function.Function;
 import java.util.stream.Gatherer;
 import java.util.stream.Stream;
 
-import static com.ginsberg.gatherers4j.GathererUtils.mustNotBeNull;
+import static com.ginsberg.gatherers4j.util.GathererUtils.mustNotBeNull;
 
 public class ZipWithGatherer<FIRST extends @Nullable Object, SECOND extends @Nullable Object>
         implements Gatherer<FIRST, Void, Pair<FIRST, SECOND>> {
