+++ src/main/java/com/ginsberg/gatherers4j/enums/package-info.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package com.ginsberg.gatherers4j;
+/// A library of useful Stream Gatherers (custom intermediate operations) for Java.
+///
+/// @author Todd Ginsberg (todd@ginsberg.com)
+@NullMarked
+package com.ginsberg.gatherers4j.enums;
 
-import java.math.BigDecimal;
-
-public record TestValueHolder(int id, BigDecimal value) {
-}
+import org.jspecify.annotations.NullMarked;
