+++ src/test/java/com/ginsberg/gatherers4j/BigDecimalProductGathererTest.java
@@ -16,6 +16,8 @@
 
 package com.ginsberg.gatherers4j;
 
+import com.ginsberg.gatherers4j.dto.WithOriginal;
+import com.ginsberg.gatherers4j.util.TestValueHolder;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
