+++ src/test/java/com/ginsberg/gatherers4j/util/TestUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Todd Ginsberg
+ * Copyright 2025 Todd Ginsberg
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,16 @@
  * limitations under the License.
  */
 
-package com.ginsberg.gatherers4j;
+package com.ginsberg.gatherers4j.util;
 
-import org.jspecify.annotations.Nullable;
+import org.assertj.core.api.recursive.comparison.RecursiveComparisonConfiguration;
 
-public record WithIndex<TYPE>(
-        long index,
-        @Nullable TYPE value
-) {
+import java.math.BigDecimal;
+
+public class TestUtils {
+
+    public static final RecursiveComparisonConfiguration BIG_DECIMAL_RECURSIVE_COMPARISON = RecursiveComparisonConfiguration
+            .builder()
+            .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
+            .build();
 }
